[feat] Logo, Notification 구조 개선
- Logo 컴포넌트 클릭 시, 메인 페이지로 이동하도록 함
- Notification의 interface 값 위치 이동

develop #3
